some stuff

+++ include/Pmr.h
@@ -4,7 +4,7 @@
 #include <stack>
 
 
-typedef std::vector<std::vector<int>> arry;
+typedef std::vector<std::vector<double>> arry;
 class Pmr {
 public:
 	Pmr();
@@ -19,7 +19,7 @@ private:
 	int n_;
 	std::vector<ArPose> points_;
 	arry graph_;
-	std::stack<int> path_;
+	std::stack<double> path_;
 
 
 	void generate_graph();
@@ -27,6 +27,6 @@ private:
 	ArPose random_point(std::vector<ArLineSegment>* lines);
 
 public:
-	std::stack<int> get_path();
+	std::stack<double> get_path();
 	std::vector<ArPose> get_points();
 };
